requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
tensorflow>=2.13.0
keras>=3.0.0
pillow>=9.0.0
Werkzeug==3.0.1
numpy>=1.21.0
reportlab>=4.0.0
scipy>=1.9.0
requests>=2.31.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.44.0
safetensors>=0.4.0
huggingface_hub>=0.17.0
python-dotenv>=1.0.0

